added test cases for orders handler and model

+++ migrations/sqls/20230906215402-product-orders-table-up.sql
@@ -1,4 +1,4 @@
 /* Replace with your SQL commands */
 
-CREATE TABLE product_orders( id SERIAL PRIMARY KEY, quantity INTEGER, order_id bigint REFERENCES orders(id) ON DELETE CASCADE,product_id bigint
+CREATE TABLE product_orders( id SERIAL PRIMARY KEY, quantity INTEGER, order_id INTEGER REFERENCES orders(id) ON DELETE CASCADE,product_id INTEGER
 REFERENCES products(id) ON DELETE CASCADE);
